Integrated optimised CausalDAG class
