+++ tests/testing_tests/test_metamorphic_relations.py
@@ -48,7 +48,7 @@ class TestMetamorphicRelation(unittest.TestCase):
         """Test if the ShouldCause MR passes all metamorphic tests where the DAG perfectly represents the program
         and there is only a single input."""
         causal_dag = CausalDAG(self.dag_dot_path)
-        causal_dag.graph.remove_nodes_from(["X2", "X3"])
+        causal_dag.remove_nodes_from(["X2", "X3"])
         adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
         should_not_cause_MR = ShouldNotCause(BaseTestCase("X1", "Z"), adj_set)
         self.assertEqual(
@@ -70,7 +70,7 @@ class TestMetamorphicRelation(unittest.TestCase):
         """Test if the ShouldCause MR passes all metamorphic tests where the DAG perfectly represents the program
         and there is only a single input."""
         causal_dag = CausalDAG(self.dag_dot_path)
-        causal_dag.graph.remove_nodes_from(["X2", "X3"])
+        causal_dag.remove_nodes_from(["X2", "X3"])
         adj_set = list(causal_dag.direct_effect_adjustment_sets(["X1"], ["Z"])[0])
         should_cause_MR = ShouldCause(BaseTestCase("X1", "Z"), adj_set)
         self.assertEqual(
@@ -218,8 +218,7 @@ class TestMetamorphicRelation(unittest.TestCase):
                 map(
                     lambda x: x.to_json_stub(skip=False),
                     filter(
-                        lambda relation: len(list(dcg.graph.predecessors(relation.base_test_case.outcome_variable)))
-                        > 0,
+                        lambda relation: len(list(dcg.predecessors(relation.base_test_case.outcome_variable))) > 0,
                         relations,
                     ),
                 )
@@ -238,8 +237,7 @@ class TestMetamorphicRelation(unittest.TestCase):
                 map(
                     lambda x: x.to_json_stub(skip=False),
                     filter(
-                        lambda relation: len(list(dag.graph.predecessors(relation.base_test_case.outcome_variable)))
-                        > 0,
+                        lambda relation: len(list(dag.predecessors(relation.base_test_case.outcome_variable))) > 0,
                         relations,
                     ),
                 )
